Refactoring to support custom Merger
GeoRocket now have a public method that returns a Merger. This method is
meant to be overriden to provide a different merger. The decision for a
Merger can be done by analyising the RoutingContext.

The Merger and the XMLMerger are now separated to be XML indepentend. The
methods: init, merge, and finishMerge; are the methods called by the
GeoRocket verticle. Implement all three methods and return an object of
the implemented Merger inside the GeoRocket.getMerger() method.

The Merger and MergerStrategy interface are moved to georocket-server-api.
