+++ georocket-server-api/src/main/java/io/georocket/output/MergeStrategy.java
@@ -1,31 +1,14 @@
 package io.georocket.output;
 
-import java.util.List;
-
 import io.georocket.storage.ChunkMeta;
 import io.georocket.storage.ChunkReadStream;
-import io.georocket.util.XMLStartElement;
 import io.vertx.core.AsyncResult;
 import io.vertx.core.Handler;
 import io.vertx.core.buffer.Buffer;
 import io.vertx.core.streams.WriteStream;
 
-/**
- * A merge strategy for XML chunks
- * @author Michel Kraemer
- */
 public interface MergeStrategy {
-  /**
-   * Set XML parent elements for the chunks to merge
-   * @param parents the parent elements
-   */
-  void setParents(List<XMLStartElement> parents);
-  
-  /**
-   * @return the merged XML parent elements
-   */
-  List<XMLStartElement> getParents();
-  
+
   /**
    * Check if a chunk with the given metadata can be merged and call a
    * handler with the result
@@ -33,7 +16,7 @@ public interface MergeStrategy {
    * @param handler will be called with the result of the operation
    */
   void canMerge(ChunkMeta meta, Handler<AsyncResult<Boolean>> handler);
-  
+
   /**
    * Initialize this merge strategy. This method must be called for all chunks that
    * should be merged. After {@link #merge(ChunkReadStream, ChunkMeta, WriteStream, Handler)}
@@ -43,7 +26,7 @@ public interface MergeStrategy {
    * the given chunk
    */
   void init(ChunkMeta meta, Handler<AsyncResult<Void>> handler);
-  
+
   /**
    * Merge an XML chunk. The given chunk should* have been passed to
    * {@link #init(ChunkMeta, Handler)} first. If it hasn't the method may or
@@ -54,12 +37,12 @@ public interface MergeStrategy {
    * @param out the stream to write the merged result to
    * @param handler will be called when the chunk has been merged
    */
-  public void merge(ChunkReadStream chunk, ChunkMeta meta, WriteStream<Buffer> out,
-      Handler<AsyncResult<Void>> handler);
-  
+  void merge(ChunkReadStream chunk, ChunkMeta meta, WriteStream<Buffer> out, Handler<AsyncResult<Void>> handler);
+
   /**
    * Finishes merging chunks and closes all open XML elements
    * @param out the stream to write the merged result to
    */
   void finishMerge(WriteStream<Buffer> out);
+
 }
